2.0 (#1)
* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* Add code styling

* wip

* Disable tests for now

* wip

* wip

+++ app/Commands/DestroyCommand.php
@@ -3,6 +3,8 @@
 namespace App\Commands;
 
 use App\Data\Sandbox;
+use App\Services\GitHub;
+use Exception;
 use LaravelZero\Framework\Commands\Command;
 
 class DestroyCommand extends Command
@@ -24,11 +26,16 @@ class DestroyCommand extends Command
     {
         $sandbox = new Sandbox;
 
+        if (! $sandbox->getSite()) {
+            throw new Exception('There is no sandbox to destroy');
+        }
+
         // Create a database backup if the site has a backup provider set
-        if (config('forge.backup_provider')) {
+        if (config('forge.enable_db') && config('forge.backup_provider')) {
             $this->components->task('Creating database backup', fn () => $sandbox->createDbBackup());
         }
 
         $this->components->task('Destroying sandbox', fn () => $sandbox->destroy());
+        $this->components->task('Posting details to GitHub', fn () => GitHub::postDestroyDetails());
     }
 }
